Be more explicit about multipart upload details
Attempts to be clearer about the handling of metadata relating to
multipart uploads which can differ between cloud providers.

The problem is that AWS S3 and Azure Blob Storage have very
different interfaces for multipart uploads. AWS S3 requres a new
multipart upload session to be created for a particular key. This
session is identified by an UploadId returned in a blob of JSON.
CloudInterface and CloudTarUploader treat this as an opaque blob,
eventually passing it to the methods _upload_part,
_complete_multipart_upload and _abort_multipart upload.

Azure Blob Storage has no such notion of a multipart upload
session. Instead blocks are uploaded with a blockid value and then
are committed by providing a list of those blockid values. This
means there is no need to track any metadata associated with the
upload other than the blockid values and the blob key.

This commit doesn't attempt to fix the underlying problem. Instead
it renames `mpu` to `upload_metadata` and updates the docstrings
so that it is clear that this is provider-specific metadata which
is treated as an opaque blob by CloudInterface.

References to the part handle are also updated in the docstrings for
similar reasons.