Enable autocommit for all postgres connections
Sets `autocommit = True` for PostgreSQLConnection.

This prevents the following situation from arising:

1. Barman starts a backup by SELECTing from pg_start_backup and
   immediately rolls back the transaction.
2. Barman starts a new transaction to retrieve the server
   version.
3. The transaction is left open putting the session into an
   idle-in-transaction state.
4. The backup lasts longer than the value of
   idle_in_transaction_session_timeout and fails because the
   session is terminated.

We used to do explicit transaction management without actually needing
transactions for anything. The only meaningful use of conn.rollback()
was to ensure that we did not leave a long-running idle transaction
after issuing pg_start_backup(), because we need to keep the session
around to issue pg_stop_backup() later.

We did not ever make changes and commit a transaction, nor did we make
changes and roll them back—in fact, other than issuing the CHECKPOINT
command or calling pg_{start,stop}_backup(), we never made any changes.
So the various calls to rollback() were an elaborate attempt to avoid
leaving idle transactions, which nevertheless did not quite succeed (in
the situation described above).

Using the context manager to avoid leaking transactions worked, but had
its own drawbacks because of changes in psycopg2 2.9, as described in
an earlier commit message. It's simpler to set autocommit=True and let
each statement be its own implicit transaction: this guarantees that we
don't leave any "idle in transaction" sessions.

We already set autocommit=True in StreamingConnection, because those
don't support transaction management anyway.

(We ought to be able to get rid of a lot of rollback handling on the
basis of this change, but this commit does not do so.)