Hold the xlogdb lock while replacing the contents of xlog.db
BackupManager.rebuild_xlogdb() would open the server.xlogdb() as fxlogdb
and then loop over the files on disk, writing an entry for each one into
fxlogdb_new. At the end, it would replace xlog.db with xlog.db.new using
shutil.move.

The problem is that it replaced the old xlog.db after the `with …` block
that held ServerXLOGDBLock, which led to the problem reported in #328:

    The archive-wal process (PID:4011354) was running. (forked from
    barman cron)

    The backup process (PID:4015392) started deleting WAL segments.
    (forked from barman backup)

    This process opened xlog.db (inode: 1613568054) and xlog.db.new
    (inode: 7058292408)

    Then the archive-wal process (PID:4011354) opened xlog.db (inode:
    1613568054) and wrote the line of 000000020002E59C000000BF to (old)
    xlog.db file.

Because we replaced xlog.db with xlog.db.new after releasing the lock,
the sequence of events above could lose valid new entries in xlog.db.

The reporter proposed to solve this problem in PR #329 by just calling
shutil.move() while holding the lock. The logic is sound, but I did not
feel comfortable mixing shutil.move with server.xlogdb's flush/fsync
operations on the original fxlogdb fd at the end of the context.

We don't write to fxlogdb during this loop, only fxlogdb_new, so there
should not be a problem in practice with using shutil.move, but I chose
to use a slightly different fix, which I found easier to follow:

1. Convert fxlogdb_new into an unnamed TemporaryFile in the same
   directory (which we don't need to worry about removing, nor
   about fsync'ing the parent directory afterwards)

2. Accumulate new entries in fxlogdb_new as before (we don't want to do
   this in memory because xlog.db could be large)

3. At the end of the loop, copy the contents of fxlogdb_new into fxlogdb
   directly using shutil.copyfileobj instead of shutil.move; this is the
   same function that .move would call, but only after reopening both
   source and destination files. (In other words, what we did before
   using shutil.move wasn't a rename-based atomic rewrite anyway.)

The same considerations apply to BackupManager.remove_wal_before_backup,
with the added wrinkle that this function used to open xlog.db readonly
and read its contents. We change that to r+ and toss in an extra seek
at the end before rewriting. We may remove lines from the file, so we
must be especially careful to truncate fxlogdb after copying data.

With this arrangement of the code, it doesn't matter that server.xlogdb
would flush/fsync the "old" (== only) fd when exiting the context. The
rewrite happens while holding the lock, so it fixes the original bug.

With thanks to KUWAZAWA Takuya for the report and original analysis.

Closes #328