Support Azure Blob Service encryption scopes
Adds the `--encryption-scope` argument to barman-cloud-backup and
barman-cloud-wal-archive. If supplied it is passed on to
AzureCloudInterface which will include it in write operations to
azure blob storage.

The value of `--encryption-scope` should be the name of an
existing encryption scope in Azure Blob Storage. When an
encryption scope is specified the key associated with that scope
is used to encrypt the data at rest in Azure Blob Storage. When
data is read from Azure Blob Storage the encryption scope is
automatically inferred and used to decrypt the data.

If the encryption scope is disabled then attempts to read data
in that scope will fail, as will attempts to write data using the
encryption scope.

Note that enabling and disabling encryption scopes within Azure
is an asynchronous operation - it will take some time (anecdotally,
several minutes) for the changes to affect the reading and writing
of data using those scopes.