Add -d option to specify database name for barman-cloud-backup.
Closes #307

By default, if a database name is not specified, it will use the
user name as the database name. When defaults (both named `postgres`)
are used, this is fine. However, if a user runs barman from a
different user, the underlying Postgres commands will fail.
To handle this case, we are adding the -d option (matching Postgres's
-d option) to allow users to specify the database name.