Add -d option to specify database name for barman-cloud-backup
barman-cloud-backup needs to connect to Postgres to start/stop the
backup, but it doesn't matter which database it connects to. If it runs
as the postgres user, it will connect to dbname=postgres by default, and
all will be well. But if it runs as another user and cannot connect to a
database of the same name, it will fail.

To handle this case, we are adding the -d option to allow users to
specify a database name explicitly.

Closes #307