Encrypt content of barman-cloud backups on s3
Passes the value of the `--encryption` argument to the call to
create_multipart_upload as the `ServerSideEncryption` parameter
so that the content of barman-cloud backup is encrypted using
the same algorithm as used for the `backup.info` file.

There is a difference in the way the `ServerSideEncryption` is
passed to `create_multipart_upload` and to `upload_fileobj`
because `upload_fileobj` expects it to be in the `ExtraArgs` dict
and `create_multipart_upload` expects it as a keyword argument.