Recreate pg_wal if required in barman-cloud-restore
If you run barman-cloud-backup on a data directory where pg_wal is a
symlink, it doesn't include pg_wal in the uploaded tar files, and the
pg_wal directory will not be created if you restore the backup.

We fix this by creating pg_wal (or pg_xlog, depending on the version) in
the destination directory if it does not exist at the end of the backup.

Closes #327