Clean up the --encryption option in barman-cloud
Removes the `--encryption` option from the following commands
because it was unused:

* barman-cloud-backup-list
* barman-cloud-restore
* barman-cloud-wal-restore

We also update the help string for the same option in the
barman-cloud-backup and barman-cloud-wal-archive commands to be
clear that the selected encryption is applied at storage time, not
transfer time.

The docs are updated to reflect the changes.