Fix pg major version in barman-cloud-backup hook script
Adds a method named pg_major_version to BackupInfo which returns
the major version of the PostgreSQL instance from which a backup
was taken.

We deliberately avoid using a property here because we do not
want pg_major_version to be included in the backup.info fields.
It is just a convenience function.

This replaces the in-line use of `int(self.version / 10000)` in
CloudBackupUploaderBarman which only produced the correct major
version for PostgreSQL >= 10.0.

This fixes an issue where barman-cloud-backup would not copy
tablespaces when running as a hook script against backups of
PostgreSQL < 10.0..