barman-cloud-wal-archive runs as pre archive hook
Updates barman-cloud-wal-archive in hook script mode such that it
supports running as a pre archive hook, not post archive.

This matches what we describe in the documentation and removes a
possible source of errors on restore due to compressing archived
WALs which have already been compressed by Barman. Such
double-compression is no longer possible as the pre archive script
runs before Barman has carried out any compression on the WAL.