Avoid leaking semaphores in barman-cloud-backup
Moves the `multiprocessing.Lock` created in `barman.copy_controller`
into the class which uses it so that the lock is not created
when the module is imported and is instead created when the object
which uses it is instantiated.

This resolves an issue where barman-cloud-backup would leak one
semaphore per worker process due to the unpickling carried out as
part of `multiprocessing.spawn`. This unpickling required
importing `barman.cloud` in the worker process which, by
dependencies, would import `barman.copy_controller` and create the
lock. The associated semaphore would then be detected by the
resource tracker resulting in warnings in the logs when running
on python >= 3.8.

Closes #362