Optimize _create_dir_and_purge for empty destination
Speeds up _create_dir_and_purge for empty destination directories
by ensuring exclude_and_protect_file is only populated if the
destination directory has content.

The motivation for this patch is: In cases where tablespaces
are not used the number of entries in exclude_and_protect_file
will be 2x the number of files in the PGDATA directory. As this
number enters the hundreds of thousands it can cause RSync to take
hours to execute the command created by _create_dir_and_purge
even if the destination directory is empty (as is the case when
a user is restoring to an empty directory, or barman is taking a
backup into an empty directory).

We can therefore speed up backup and restore times in these cases
by not having RSync apply the filter rules in
exclude_and_protect_file.

This is achieved by updating _analyze_directory so that it
only populates exclude_and_protect_file if the destination
directory has content. This is determined by setting
ref_has_content to True if _list_files returns *anything* which
isn't the current directory. This meant doing away with the dict
comprehension and replacing it with a for loop since the presence
of a directory counts as content. We therefore must check for
directories - if we filter them out in a comprehension then we
would have to do another pass over the output of _list_files in
order to determine whether there is any content.

Closes #378