Refactor: Move ALLOWED_COMPRESSIONS to barman.cloud
Moves ALLOWED_COMPRESSIONS so that it can also be used in cloud.py
without creating a circular dependency.

This is groundwork for #368 where a barman-cloud-delete command
will be added to barman-cloud - we will need to detect WAL files
which could have no compression suffix or a suffix for any of the
allowed compressions.