Refactor: Allow for `RetentionPolicy` re-use
The `retention_policy` module is refactored so that the
`RetentionPolicy` classes can be re-used by barman-cloud. Without
this patch a `RetentionPolicy` can only be used with `server.Server`
which requires access to a barman server configuration (something
that the barman-cloud scripts do not have).

This is achieved by making it possible to create a `RetentionPolicy`
from a server name and pre-generated list of `BackupInfo` objects,
as an alternative to just passing in a `server.Server` object.

A new `ServerMetadata` class is created to encapsulate the server
data which the retention policies need to access. This is the
server name, the minimum redundancy value, any existing
`RetentionPolicy` object in the config (needed by
`SimpleWALRetentionPolicy`) and a list of available backups.

The factory method handles the details of the `ServerMetadata` and
will create the `ServerMetadata` from either a supplied
`barman.Server` object or a server name and `BackupInfo` list.

The existing barman code which uses retention policies will
continue to behave in the same way, that is the available backups
will be generated each time it is called and the server metadata
will be obtained from the server config.

barman-cloud will be able to pass the name and a `BackupInfo` list
(obtained from a `CloudBackupCatalog`) and create a "static"
`RetentionPolicy` which will always report the same data.

Because barman-cloud does not have access to a barman.config, the
minimum_redundancy for the server is set to 0.

This is groundwork for #368 where a barman-cloud-delete command
will be added to barman-cloud - we will need to be able to apply
the rention policy logic to backups stored in the cloud.