Refactor: Expose `BackupManager` logic for re-use
Extracts core logic from `get_next_backup` and `get_previous_backup`
into static methods which are then called by the instance methods.
This is so that barman-cloud, which uses the barman code as a
library, can use that same logic by passing in its own dict of
`BackupInfo` objects - it will not have to create its own
`BackupManager` instance.

For the same reason, we also extract the logic which finds timelines
to protect into a static method and call it from `delete_backup`
method.

This isn't particularly expressive or meaningful - it's literally
just separating the stateless bits of the logic which can be
applied to state barman-cloud will pass in and adapting
`BackupManager` to pass its own state to those static methods in the
same way.

This is groundwork for #368 where a barman-cloud-delete command
will be added to barman-cloud - we will need to be able to find
the previous backup to the backup being deleted in order to
determine whether any WALs can be removed. We will also need to use
the timeline protection logic to ensure that we do not remove WALs
from timelines unrelated to the backup being deleted.