Implement bulk object deletion in cloud interfaces
Adds delete_objects method to CloudInterface and its two
implementations.

For both S3CloudInterface and AzureCloudInterface we use the
functions provided by the respective SDKs.

In boto3, the delete_objects function has a documented limit of
1000 keys so we batch requests accordingly to avoid exceeding this
limit. We use Quiet=True to limit the amount of information about
each deleted object to errors only.

The Azure SDK does not have a documented limit for its delete_blobs
function so we do not batch anything. There is also no analog for
Quiet=True.

For both cloud providers we check the response for errors and log
the affected objects before raising a CloudProviderError.

This is groundwork for #368 where a barman-cloud-delete command
will be added to barman-cloud - we will need to be able to delete
the files associated with backups as well as any related WALs.