Track unreadable backups in `CloudBackupCatalog`
Updates `CloudBackupCatalog` so that it keeps a record of the
backups for which it was unable to read the `backup.info` file.

This is groundwork for #368 where a barman-cloud-delete command
will be added to barman-cloud - this command will need to know if
there were any backups that could not be listed as it is not safe
to make decisions around retention policy or WAL cleanup if we do
not have information for every backup in cloud storage.