Enhance `CloudBackupCatalog` so it can find WALs
Adds the method `CloudBackupCatalog.get_wal_paths` which retrieves
the full path of every WAL in cloud storage keyed by the name of
the WAL.

This is implemented with the same caching behaviour used by
`get_backup_list`.

This is groundwork for #368 where a barman-cloud-delete command
will be added to barman-cloud - this command will need to be able
to view all WALs in cloud storage in order to determine which
WALs are no longer required once a backup has been removed.