Support removal of cached items from `CloudBackupCatalog`
Adds the methods `remove_backup_from_cache` and `remove_wal_from_cache`
to `CloudBackupCatalog which will remove the supplied key from the
list of cached backups or WAL paths respectively.

This is groundwork for #368 where a barman-cloud-delete command
will be added to barman-cloud - when deleting multiple backups
and WALs it will be necessary to remove deleted backups and WALs
from the `CloudBackupCatalog` cache without requiring a scan of the
remote bucket.