Optionally allow missing files in `CloudBackupCatalog`
Adds the allow_missing parameter to `CloudBackupCatalog.get_backup_files`
which allows the original behaviour of raising a SystemExit to be
overridden and the backup files with missing paths to be returned with
a `path` of `None` instead.

Some rudimentary tests are added for the `get_backup_files` method,
along with a specific test to verify the `allow_missing=True` behaviour.

This is groundwork for #368 where a barman-cloud-delete command will be
added to barman-cloud - this command will need to be able to proceed
even if some files could not be found, in order to handle cases where
previous deletions have failed partially and removed only some of the
files. In such cases the original behaviour of raising SystemExit is not
appropriate.