Implement barman-cloud-backup-delete
Adds the command barman-cloud-backup-delete which can delete backups in
cloud storage and remove any related WALs.

Backups can be deleted either by backup ID (-b/--backup-id) or by
retention policy (-r/--retention-policy). The supported retention
policies are exactly the same as those for barman-server.

This implementation makes some opinionated choices:

1. Backups are deleted one at a time

Instead of building a list of all objects to delete for all backups and
passing them through to the cloud provider, we delete each backup
individually.

For large numbers of backups this is significantly less efficient (and
more expensive) however in practice the amount of backups to be deleted
due to a retention policy will be at most 1, assuming that
barman-cloud-backup-delete runs frequently enough. Optimising for the
case where a large number of backups have accumulated doesn't seem worth
it given the complications it introduces (namely figuring out the
affected backups in the event of a partial failure of a batch delete,
and determining the WALs to clean up after the completion of the batch,
taking any failed deletions into account).

2. WALs are removed after successful deletion of each backup

For similar reasons to 1, we remove the WALs for each backup after it
has been successfully deleted rather than waiting until all backups have
been deleted and performing a final cleanup.

3. Backup files and WALs are explicitly sorted before being deleted

Although the ordering makes little difference to the cloud provider
there are benefits to logging in a specific and deterministic sort
order, which is not guaranteed with dicts in python <3.6.

4. We copy the behaviour of `backup.BackupManager` for WAL cleanup

We use the same behaviour implemented in `backup.BackupManager` when
finding WALs to clean up. This leads to two consequences:

* If there is an older backup on an unrelated timeline we will not clean
  up WALs.
* If there are any WALs left on a timeline with no backups we will clean
  them up.

I'm not 100% clear on whether either of these things are what we want -
there are tests which verify the behaviours currently tagged with TODO.

5. We only remove WALs *up to* the begin_wal of the deleted backup

Based on comments in `backup.BackupManager`, it is only safe to remove
WALs up to the begin_wal of the oldest remaining backup after deletion
if exclusive backups are used. If concurrent (non-exclusive) backups
are used then it is possible the WALs associated with the deleted
backup could be useful to other backups started just after the deleted
backup. Because exclusive backups are deprecated we must assume
non-exclusive backups are used and therefore we can only remove WALs
prior to the deleted backup.

Because the backup label is associated with the deleted backup it
will not be deleted until the next backup is deleted, at which point
it will be cleaned up with the rest of the WALs prior to that backup.

6. We exit immediately if an error is encountered on deletion

The benefit of this approach is it forces the user to resolve whatever
issue is preventing the deletion of a backup as they simply won't be
able to delete by retention policy until the offending backup is
deleted.

7. We delete backup.info separately after deleting the rest of the files

Rather than include `backup.info` in the `delete_objects` request we
delete it in a separate request after successfully deleting the rest of
the backup files.

This means that if there are any failures deleting backup files we will
still have the `backup.info` file so will be able to try again.

8. We silently ignore files which are expected to exist but do not

This allows us to successfully retry from a partial failure because any
missing files are considered to have already been deleted. This is
consistent with the way most cloud object stores handle deletes on
missing files.