Clarify comments around WAL cleanup
Updates the comments around the code which determines which backup
is used as the cut off point for removing WALs because they assumed
some implicit knowledge which the reader may not have.

The re-wording is based on the original comments, last re-written
in 87754a704668a45b8524c2ffbfc36a5b4a52dce8.