Move pg_receieve{wal,xlog} after the connection check
Moves the server version check for determining whether we log
pg_receivewal or pg_receivexlog so that it occurs after the initial
connection check.

This fixes the "Barman using replication slots server unreachable fails
running receive-wal" integration test which was breaking because the
call to self.server.streaming.server_version was raising an exception
directly and the expected log messages were not being logged.