Add support for Azure SAS token at blob container level
Replaces the implementation of AzureCloudInterface._check_bucket_existence
so that it does not use ContainerClient.exists. Instead we use
ContainerClient.list_blobs as a proxy for existence - if we can call
next() on the return value of list_blobs and trigger either no exception
or a StopIteration then we know the bucket exists.

This fixes the only issue which prevented the azure blob storage cloud
provider from working with SAS tokens created at the blob container
level, specifically that calling ContainerClient.exists requires storage
account permissions, not container permissions.

We also change the logic around creating the ContainerClient so that we
instantiate it directly rather than first creating a BlobServiceClient.
This isn't strictly required to fix the SAS token issue but since it
results in a simpler connection initialisation it is worth doing anyway.

Closes #388