Forward config path from passive to primary node
Adds a new config option for passive nodes named `forward_config_path`.
If this is set to True then the location of the barman config on the
passive node filesystem will be used as the `-c/--config` option when
invoking `barman sync-info` on the primary node.

The default is False in order to preserve the existing behaviour which
does not apply the `-c/--config` option from the passive node when
invoking `barman sync-info` on the primary node. This is something we
should consider changing on our next major version number release.

Note: The remote `barman sync-info` invocation in
`server.Server.primary_node_info` is the only place where the passive
node needs to run `barman` on the primary node. When this runs as part
of `barman cron` it persists the sync-info to a local file named
`primary.info` - any other passive barman commands such as `sync-wals`
and `sync-backup` will then use that `primary.info` file directly.

Closes #371