Add support for backup annotations
Allows backups to have annotations. These are key/value pairs where the
key is either a file name (in barman) or an object key (in barman cloud)
and the value is the contents of the file/object.

This is implemented through AnnotationManagerFile and
AnnotationManagerCloud which provide an interface defined in the
AnnotationManager abstract base class.

This commit just adds the AnnotationManager and associated tests.
Annotations themselves are not (yet) intended to be exposed to users
through the CLI.

Barman itself is not updated to care about annotations. Any behavioural
changes should be implemented along with the feature being built on top
of annotations.

Note that AnnotationmanagerCloud optimizes for the case of sparse
annotations by building a cache of annotation existence at the cost of
one request to list bucket keys. This will save on remote_open requests
in cases where we need to look for annotations across multiple backups
(e.g. when listing backups or when applying retention policies). For
cases where we only care about one backup we can optionally skip
building the cache.

The is groundwork for #370 where we will add archival backup
functionality which will be implemented using an annotation named
"keep".