barman keep: Add KeepManager
Adds the functinoality to create, read and delete the keep status
of a backup. The keep status is tracked using annotations as
follows:

  - An archival backup has an annotation named `keep`.
  - The value of that annotation is the intended recovery target
    which can be either "standalone" (no recovery from that backup
    to a future point in time) or "full" (recovery is possible from
    the archival backup to the present).

Barman code can interact with the keep status using the KeepManager
which will transparently use either a filesystem or a cloud backend
depending on how it was instantiated.

The keep functionality is not exposed via the CLI at this point.

The is groundwork for adding archival backup functionality in #370
which will be implemented using the KeepManager to create, read and
delete the keep status of backups.