barman keep: Add `barman keep` command
Adds a fully functional `barman keep` command which allows keep
annotations to be added and removed from backups and for the value
to be inspected.

The recovery target specified by `--target` is required and deliberately
has no default. The intended default was "standalone" however applying
a default target can lead to potentially confusing behaviour, e.g.:
Running a `barman keep` command with no `--target` against a backup
which already has a "full" recovery target would change it from "full"
to "standalone". This could result in WALs subsequently removed to the
surprise of the barman user. We therefore always require the user to
specify the intended recovery target.

Barman is not yet modified to respect the keep value - this will happen
in subsequent commits.

This is part of the implementation for #370.