barman keep: Delete by backup ID in barman-cloud
Updates the implementation of `barman-cloud-delete-backup` so that it
will not delete an archival backup, i.e. a backup with a keep
annotation.

This changes the behaviour of the `barman-cloud-backup-delete` command
such that a user cannot delete a backup which has a keep annotation -
instead the command fails and they get a helpful error message.

This is part of the implementation for #370.