barman keep: Handle keep in retention policy for barman-cloud
Updates barman-cloud so that it can use the updated RetentionPolicy
logic which is aware of archival backups. This is just a case of
making sure barman-cloud is able to provide a KeepManager to the
RetentionPolicy as the logic in the retention policies is already
implemented.

This is part of the implementation for #370.