barman keep: Handle keep in retention policy for barman
Updates retention polocies to be aware of the keep annotation.
RedundancyRetentionPolicy and RecoveryWindowRetentionPolicy are updated
so that they report the value of the keep recovery target instead of the
usual VALID, OBSOLETE or POTENTIALLY_OBSOLETE values.

This has the following consequences:

  - The list-backup command now reports the keep recovery target.
  - The cron functionality will not delete backups which have a keep
    recovery target.

Although the BackupManager.cron_retention_policy function is already
implemented such that only OBSOLETE backups are deleted, a test is added
in this commit so that the logic which preserves non-OBSOLETE backups is
explicitly tested.

This is part of the implementation for #370.