barman keep: Delete by backup ID in barman
Updates the implementation of delete_backup so that it will not delete
an archival backup, i.e. a backup with a keep annotation.

This changes the behaviour of the `barman delete` command such that
a user cannot delete a backup which has a keep annotation - instead
the command fails and they get a helpful error message.

This is part of the implementation for #370.