barman keep: Test WAL cleanup behaviour
Adds tests which explicitly test the behaviour of wal cleanup logic when
backups are deleted. The tests are checking existing functionality so no
code changes are required.

These tests are being added because the wal cleanup logic will need to
change and we need to be especially careful that we do not break any of
the existing behaviour.

This is part of the implementation for #370.