barman keep: Make barman WAL cleanup keep-aware
Makes the code which handles cleanup of WALs when a backup is
removed aware of archival backups.

The old logic of checking whether there is a previous backup and
cleaning up WALs if there isn't is replaced with a function called
`should_remove_wals` which determines:

  1. Whether WALs can be removed during the deletion of this backup,
     depending on whether there are any previous backups and whether
     those backups are standalone archival backups (the logic itself
     is described in the docstring for `should_remove_wals`).
  2. Whether there are any WAL ranges which must not be deleted
     because they are needed by standalone archival backups.

Once this has been determined, the rest of WAL cleanup is simply a
matter of checking none of the WALs are within the protected WAL ranges,
in addition to the existing checks.

One complication is the presence of the rather awkwardly named
`skip_wal_cleanup_if_standalone` parameter which is used to skip WAL
cleanup whenever the only previous backups are standalone archival
backups. The reason for this is because if we always clean up WALs for
standalone archival backups when a backup is deleted we risk the
following scenario:

  1. A user immediately tags a new backup as keep:standalone and they
     have no older backups.
  2. They take a new backup but then change their mind and delete it.
     The WALs after the end_wal of the keep:standalone backup will get
     cleaned up, leaving the user unable to recover to the present from
     that backup.

We therefore only clean up WALs in the "all older backups are standalone
archival backups" case when deleting by retention policy, because the
retention policy will have determined the backups to be obsolete before
we get to cleaning up WALs.

This clean up is enabled by the `cron_retention_policy` method calling
`delete_backup` with `skip_wal_cleanup_if_standalone=False`.

This is part of the implementation of #370.